requirements.txt
asgiref==3.8.1
Django==4.2.14
django-allauth==0.63.6
django-countries==7.6.1
django-humanize==0.1.2
django-scheduler==0.10.1
django-simple-history==3.7.0
django-tinymce==4.1.0
fontawesomefree==6.6.0
humanize==4.10.0
icalendar==5.0.13
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
gunicorn
psycopg2-binary
